Added upload_file and check whether bucket exists

+++ src/aws/resourcemanager/resourcemanager.py
@@ -14,13 +14,15 @@ class ResourceManagerCore:
 
     def __init__(self):
         self.S3 = boto3.client('s3')
+        self.S3_resource = boto3.resource('s3')
 
     def run(self):
-        try:
-            bucket_name, bucket_response = self.create_bucket()
-            self.delete_bucket(bucket_name)
-        except ClientError:
-            print("You should add the AmazonS3ReadOnlyAccess and AmazonS3FullAccess permission to the user")
+        # try:
+        bucket_name, bucket_response = self.create_bucket()
+        self.delete_bucket("hallo")
+
+        # except ClientError:
+        #     print("You should add the AmazonS3ReadOnlyAccess and AmazonS3FullAccess permission to the user")
         print(self.S3.list_buckets())
 
     def create_bucket(self):
@@ -39,10 +41,27 @@ class ResourceManagerCore:
 
     def delete_bucket(self, bucket_name):
         """
-        Method called to delete the bucket with the name bucket_name.
+        Method called to delete the bucket with the name 'bucket_name'.
         :param bucket_name: Name of the bucket to be deleted.
         """
-        self.S3.delete_bucket(Bucket=bucket_name)
+        if self.S3_resource.Bucket(bucket_name).creation_date is None:
+            print("Bucket " + bucket_name + " does not exist")
+        else:
+            self.S3.delete_bucket(Bucket=bucket_name)
+
+    def upload_file(self, file_path, bucket_name, key):
+        """
+        Method called to upload a file with path 'file_path' to the bucket with
+        name 'bucket_name' and upload it to the storage with name 'key'.
+        :param file_name: Path of the file to be uploaded.
+        :param bucket_name: Name of the bucket to upload the file to.
+        :param key: Name of the key to upload to.
+        """
+        if self.S3_resource.Bucket(bucket_name).creation_date is None:
+            print("Bucket " + bucket_name + " does not exist")
+        else:
+            self.S3.upload_file(file_path, bucket_name, key)
+
 
 class ResourceMonitor(Listener):
